move NServiceBusStartup classes into Startup folders/namespace, as suggested in review
