+++ src/SFA.DAS.EmployerFinance.MessageHandlers/Startup/NServiceBusStartup.cs
@@ -10,7 +10,7 @@ using SFA.DAS.NServiceBus.NLog;
 using SFA.DAS.NServiceBus.StructureMap;
 using StructureMap;
 
-namespace SFA.DAS.EmployerFinance.MessageHandlers
+namespace SFA.DAS.EmployerFinance.MessageHandlers.Startup
 {
     public class NServiceBusStartup : IRunAtStartup
     {
